refactor(slack): split message dispatch
